+++ src/main/java/com/jsk/schedule/domain/notification/entity/NotificationChannel.java
@@ -1,5 +1,0 @@
-package com.jsk.schedule.domain.notification.entity;
-
-public enum NotificationChannel {
-    KAKAO
-}