+++ src/main/java/com/jsk/schedule/domain/notification/entity/NotificationType.java
@@ -1,5 +1,0 @@
-package com.jsk.schedule.domain.notification.entity;
-
-public enum NotificationType {
-    SCHEDULE_CREATED, SCHEDULE_UPDATED, SCHEDULE_DELETED
-}