+++ src/main/java/com/jsk/schedule/domain/schedule/entity/ScheduleType.java
@@ -1,5 +1,0 @@
-package com.jsk.schedule.domain.schedule.entity;
-
-public enum ScheduleType {
-    VACATION, WORK
-}